requirements.txt
altair==5.1.1
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
bqplot==0.12.40
branca==0.6.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
colour==0.1.5
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.4
Fiona==1.9.4.post1
folium==0.14.0
fonttools==4.42.1
gdown==4.7.1
geojson==3.0.1
geopandas==0.14.0
gitdb==4.0.10
GitPython==3.1.36
idna==3.4
importlib-metadata==6.8.0
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.17.4
ipython==8.15.0
ipytree==0.2.2
ipywidgets==8.1.1
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.17.3
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
leafmap==0.25.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
numpy==1.25.2
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.24.3
pure-eval==0.2.2
pyarrow==13.0.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.1
pyproj==3.6.0
pyrsistent==0.19.3
pyshp==2.3.1
PySocks==1.7.1
pystac==1.8.3
pystac-client==0.7.5
python-box==7.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
rich==13.5.2
scooby==0.7.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
soupsieve==2.5
stack-data==0.6.2
streamlit==1.26.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
traittypes==0.2.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==4.0.9
xyzservices==2023.7.0
zipp==3.16.2
